+++ src/mcp-server/server.ts
@@ -28,7 +28,6 @@ import { tool$userCreateUser } from "./tools/userCreateUser.js";
 import { tool$userCreateUsersWithListInput } from "./tools/userCreateUsersWithListInput.js";
 import { tool$userDeleteUser } from "./tools/userDeleteUser.js";
 import { tool$userGetUserByName } from "./tools/userGetUserByName.js";
-import { tool$userLoginUser } from "./tools/userLoginUser.js";
 import { tool$userLogoutUser } from "./tools/userLogoutUser.js";
 import { tool$userUpdateUser } from "./tools/userUpdateUser.js";
 
@@ -43,7 +42,7 @@ export function createMCPServer(deps: {
 }) {
   const server = new McpServer({
     name: "Petstore",
-    version: "5.2.0-alpha.23",
+    version: "5.2.0-alpha.25",
   });
 
   const client = new PetstoreCore({
@@ -87,7 +86,6 @@ export function createMCPServer(deps: {
   tool(tool$storeDeleteOrder);
   tool(tool$userCreateUser);
   tool(tool$userCreateUsersWithListInput);
-  tool(tool$userLoginUser);
   tool(tool$userLogoutUser);
   tool(tool$userGetUserByName);
   tool(tool$userUpdateUser);