+++ src/models/operations/index.ts
@@ -10,6 +10,5 @@ export * from "./findpetsbytags.js";
 export * from "./getorderbyid.js";
 export * from "./getpetbyids.js";
 export * from "./getuserbyname.js";
-export * from "./loginuser.js";
 export * from "./updateuser.js";
 export * from "./uploadfile.js";