+++ src/sdk/user.ts
@@ -6,7 +6,6 @@ import { userCreateUser } from "../funcs/userCreateUser.js";
 import { userCreateUsersWithListInput } from "../funcs/userCreateUsersWithListInput.js";
 import { userDeleteUser } from "../funcs/userDeleteUser.js";
 import { userGetUserByName } from "../funcs/userGetUserByName.js";
-import { userLoginUser } from "../funcs/userLoginUser.js";
 import { userLogoutUser } from "../funcs/userLogoutUser.js";
 import { userUpdateUser } from "../funcs/userUpdateUser.js";
 import { ClientSDK, RequestOptions } from "../lib/sdks.js";
@@ -43,20 +42,6 @@ export class User extends ClientSDK {
     options?: RequestOptions,
   ): Promise<components.User> {
     return unwrapAsync(userCreateUsersWithListInput(
-      this,
-      request,
-      options,
-    ));
-  }
-
-  /**
-   * Logs user into the system
-   */
-  async loginUser(
-    request: operations.LoginUserRequest,
-    options?: RequestOptions,
-  ): Promise<operations.LoginUserResponse> {
-    return unwrapAsync(userLoginUser(
       this,
       request,
       options,