+++ RELEASES.md
@@ -603,3 +603,13 @@ Based on:
 - [typescript v5.2.0-alpha.23] .
 ### Releases
 - [NPM v5.2.0-alpha.23] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.23 - .
+
+## 2025-08-05 22:47:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.519.0 (2.552.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.25] .
+### Releases
+- [NPM v5.2.0-alpha.25] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.25 - .